+++ crates/core/src/lib.rs
@@ -12,7 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#[cfg(feature = "r0vm")]
 mod crypto;
 
 use alloy_primitives::{Address, B256, Bytes, KECCAK256_EMPTY, U256, keccak256, map::B256Map};
@@ -26,8 +25,7 @@ use reth_trie_common::{EMPTY_ROOT_HASH, HashedPostState, TrieAccount};
 use risc0_ethereum_trie::CachedTrie;
 use std::{cell::RefCell, collections::hash_map::Entry, fmt::Debug, marker::PhantomData};
 
-#[cfg(feature = "r0vm")]
-pub use crypto::{R0vmCrypto, install_r0vm_crypto};
+pub use crypto::install_r0vm_crypto;
 pub use reth_stateless::{ExecutionWitness, StatelessTrie, UncompressedPublicKey};
 
 pub type EthEvmConfig<C> = reth_evm_ethereum::EthEvmConfig<C, EthEvmFactory>;
@@ -71,7 +69,7 @@ where
         "only post-merge blocks supported"
     );
 
-    #[cfg(all(feature = "r0vm", target_os = "zkvm", target_vendor = "risc0"))]
+    #[cfg(all(target_os = "zkvm", target_vendor = "risc0"))]
     assert!(install_r0vm_crypto());
 
     let (hash, _) = reth_stateless::stateless_validation_with_trie::<SparseState, _, _>(